Refactor OutputLayer into BaseOutputLayer, add RnnOutputLayer, modify RNN tests
