+++ deeplearning4j-core/src/main/java/org/deeplearning4j/nn/layers/factory/DefaultLayerFactory.java
@@ -81,6 +81,8 @@ public class DefaultLayerFactory implements LayerFactory {
         	return new org.deeplearning4j.nn.layers.recurrent.GRU(conf);
         if(layerConfig instanceof org.deeplearning4j.nn.conf.layers.OutputLayer)
             return new org.deeplearning4j.nn.layers.OutputLayer(conf);
+        if(layerConfig instanceof org.deeplearning4j.nn.conf.layers.RnnOutputLayer)
+        	return new org.deeplearning4j.nn.layers.recurrent.RnnOutputLayer(conf);
         if(layerConfig instanceof org.deeplearning4j.nn.conf.layers.RecursiveAutoEncoder)
             return new org.deeplearning4j.nn.layers.feedforward.autoencoder.recursive.RecursiveAutoEncoder(conf);   
         if(layerConfig instanceof org.deeplearning4j.nn.conf.layers.ConvolutionLayer)
