+++ deeplearning4j-core/src/test/java/org/deeplearning4j/gradientcheck/GradientCheckTests.java
@@ -14,6 +14,7 @@ import org.deeplearning4j.nn.conf.layers.DenseLayer;
 import org.deeplearning4j.nn.conf.layers.GRU;
 import org.deeplearning4j.nn.conf.layers.GravesLSTM;
 import org.deeplearning4j.nn.conf.layers.OutputLayer;
+import org.deeplearning4j.nn.conf.layers.RnnOutputLayer;
 import org.deeplearning4j.nn.conf.preprocessor.RnnToFeedForwardPreProcessor;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
 import org.deeplearning4j.nn.weights.WeightInit;
@@ -228,10 +229,9 @@ public class GradientCheckTests {
 	        .layer(1, new GRU.Builder().nIn(gruLayerSize).nOut(gruLayerSize).activation("tanh")
 	        		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1.0))
 	        		.updater(Updater.NONE).build())
-	        .layer(2, new OutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(gruLayerSize)
+	        .layer(2, new RnnOutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(gruLayerSize)
 	        		.nOut(nOut).weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1.0))
 	        		.updater(Updater.NONE).build())
-	        .inputPreProcessor(2, new RnnToFeedForwardPreProcessor())
 	        .pretrain(false).backprop(true)
 	        .build();
     	
@@ -247,10 +247,12 @@ public class GradientCheckTests {
     			}
     		}
     	}
-    	INDArray labels = Nd4j.zeros(miniBatchSize*timeSeriesLength,nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-    	for( int i=0; i<labels.size(0); i++){
-    		int idx = r.nextInt(nOut);
-    		labels.putScalar(new int[]{i,idx}, 1.0f);
+    	INDArray labels = Nd4j.zeros(miniBatchSize,nOut,timeSeriesLength);
+    	for( int i=0; i<miniBatchSize; i++ ){
+    		for( int j=0; j<timeSeriesLength; j++ ){
+    			int idx = r.nextInt(nOut);
+    			labels.putScalar(new int[]{i,idx,j}, 1.0f);
+    		}
     	}
     	
     	if( PRINT_RESULTS ){
@@ -286,10 +288,12 @@ public class GradientCheckTests {
     			}
     		}
     	}
-    	INDArray labels = Nd4j.zeros(miniBatchSize*timeSeriesLength,nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-    	for( int i=0; i<labels.size(0); i++){
-    		int idx = r.nextInt(nOut);
-    		labels.putScalar(new int[]{i,idx}, 1.0f);
+    	INDArray labels = Nd4j.zeros(miniBatchSize,nOut,timeSeriesLength);
+    	for( int i=0; i<miniBatchSize; i++ ){
+    		for( int j=0; j<timeSeriesLength; j++ ){
+    			int idx = r.nextInt(nOut);
+    			labels.putScalar(new int[]{i,idx,j}, 1.0f);
+    		}
     	}
         
         double[] l2vals = {0.0, 0.4, 0.0};
@@ -311,10 +315,9 @@ public class GradientCheckTests {
 			                .layer(0, new GRU.Builder().nIn(nIn).nOut(layerSize).activation(afn)
 			                		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1))
 			                		.updater(Updater.NONE).build())
-			                .layer(1, new OutputLayer.Builder(lf).activation(outputActivation).nIn(layerSize).nOut(nOut)
+			                .layer(1, new RnnOutputLayer.Builder(lf).activation(outputActivation).nIn(layerSize).nOut(nOut)
 			                		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1))
 			                		.updater(Updater.NONE).build())
-			                .inputPreProcessor(1, new RnnToFeedForwardPreProcessor())
 			                .pretrain(false).backprop(true)
 			                .build();
 			
@@ -359,11 +362,13 @@ public class GradientCheckTests {
         			}
         		}
         	}
-        	
-        	INDArray labels = Nd4j.zeros(miniBatchSize[i]*timeSeriesLength[i],nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-        	for( int m=0; m<labels.size(0); m++){
-        		int idx = r.nextInt(nOut);
-        		labels.putScalar(new int[]{m,idx}, 1.0f);
+
+        	INDArray labels = Nd4j.zeros(miniBatchSize[i],nOut,timeSeriesLength[i]);
+        	for( int m=0; m<miniBatchSize[i]; m++ ){
+        		for( int j=0; j<timeSeriesLength[i]; j++ ){
+        			int idx = r.nextInt(nOut);
+        			labels.putScalar(new int[]{m,idx,j}, 1.0f);
+        		}
         	}
     		
     		MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder()
@@ -372,9 +377,8 @@ public class GradientCheckTests {
 	            .list(2)
 	            .layer(0, new GRU.Builder().nIn(nIn).nOut(layerSize)
 	            		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1)).updater(Updater.NONE).build())
-	            .layer(1, new OutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
+	            .layer(1, new RnnOutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
 	            		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1)).updater(Updater.NONE).build())
-	            .inputPreProcessor(1, new RnnToFeedForwardPreProcessor())
 	            .pretrain(false).backprop(true)
 	            .build();
     		MultiLayerNetwork mln = new MultiLayerNetwork(conf);
@@ -407,9 +411,8 @@ public class GradientCheckTests {
 	        		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1.0)).updater(Updater.NONE).build())
 	        .layer(1, new GravesLSTM.Builder().nIn(layerSize).nOut(layerSize).activation("tanh")
 	        		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1.0)).updater(Updater.NONE).build())
-	        .layer(2, new OutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
+	        .layer(2, new RnnOutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
 	        		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1.0)).updater(Updater.NONE).build())
-	        .inputPreProcessor(2, new RnnToFeedForwardPreProcessor())
 	        .pretrain(false).backprop(true)
 	        .build();
     	
@@ -425,10 +428,12 @@ public class GradientCheckTests {
     			}
     		}
     	}
-    	INDArray labels = Nd4j.zeros(miniBatchSize*timeSeriesLength,nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-    	for( int i=0; i<labels.size(0); i++){
-    		int idx = r.nextInt(nOut);
-    		labels.putScalar(new int[]{i,idx}, 1.0f);
+    	INDArray labels = Nd4j.zeros(miniBatchSize,nOut,timeSeriesLength);
+    	for( int i=0; i<miniBatchSize; i++ ){
+    		for( int j=0; j<timeSeriesLength; j++ ){
+    			int idx = r.nextInt(nOut);
+    			labels.putScalar(new int[]{i,idx,j}, 1.0f);
+    		}
     	}
     	
     	if( PRINT_RESULTS ){
@@ -464,11 +469,15 @@ public class GradientCheckTests {
     			}
     		}
     	}
-    	INDArray labels = Nd4j.zeros(miniBatchSize*timeSeriesLength,nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-    	for( int i=0; i<labels.size(0); i++){
-    		int idx = r.nextInt(nOut);
-    		labels.putScalar(new int[]{i,idx}, 1.0f);
+    	
+    	INDArray labels = Nd4j.zeros(miniBatchSize,nOut,timeSeriesLength);
+    	for( int i=0; i<miniBatchSize; i++ ){
+    		for( int j=0; j<timeSeriesLength; j++ ){
+    			int idx = r.nextInt(nOut);
+    			labels.putScalar(new int[]{i,idx,j}, 1.0f);
+    		}
     	}
+    	
         
         double[] l2vals = {0.0, 0.4, 0.0};
         double[] l1vals = {0.0, 0.0, 0.5};	//i.e., use l2vals[i] with l1vals[i]
@@ -489,10 +498,9 @@ public class GradientCheckTests {
 			                .layer(0, new GravesLSTM.Builder().nIn(nIn).nOut(layerSize)
 			                		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1))
 			                		.activation(afn).updater(Updater.NONE).build())
-			                .layer(1, new OutputLayer.Builder(lf).activation(outputActivation).nIn(layerSize).nOut(nOut)
+			                .layer(1, new RnnOutputLayer.Builder(lf).activation(outputActivation).nIn(layerSize).nOut(nOut)
 			                		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1))
 			                		.updater(Updater.NONE).build())
-			                .inputPreProcessor(1, new RnnToFeedForwardPreProcessor())
 			                .pretrain(false).backprop(true)
 			                .build();
 			
@@ -538,10 +546,12 @@ public class GradientCheckTests {
         		}
         	}
         	
-        	INDArray labels = Nd4j.zeros(miniBatchSize[i]*timeSeriesLength[i],nOut);	//Would be this shape after reshaping 3d -> 2d for output layer
-        	for( int m=0; m<labels.size(0); m++){
-        		int idx = r.nextInt(nOut);
-        		labels.putScalar(new int[]{m,idx}, 1.0f);
+        	INDArray labels = Nd4j.zeros(miniBatchSize[i],nOut,timeSeriesLength[i]);
+        	for( int m=0; m<miniBatchSize[i]; m++){
+        		for( int j=0; j<timeSeriesLength[i]; j++ ){
+        			int idx = r.nextInt(nOut);
+        			labels.putScalar(new int[]{m,idx,j}, 1.0f);
+        		}
         	}
     		
     		MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder()
@@ -550,9 +560,8 @@ public class GradientCheckTests {
 	            .list(2)
 	            .layer(0, new GravesLSTM.Builder().nIn(nIn).nOut(layerSize).weightInit(WeightInit.DISTRIBUTION)
 	            		.dist(new NormalDistribution(0,1)).updater(Updater.NONE).build())
-	            .layer(1, new OutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
+	            .layer(1, new RnnOutputLayer.Builder(LossFunction.MCXENT).activation("softmax").nIn(layerSize).nOut(nOut)
 	            		.weightInit(WeightInit.DISTRIBUTION).dist(new NormalDistribution(0,1)).updater(Updater.NONE).build())
-	            .inputPreProcessor(1, new RnnToFeedForwardPreProcessor())
 	            .pretrain(false).backprop(true)
 	            .build();
     		MultiLayerNetwork mln = new MultiLayerNetwork(conf);
